Fix memory issues when using Vector to draw a grid

+++ include/2DEngine/shapes/Grid.h
@@ -22,7 +22,7 @@ class Grid: public Shape {
         GETTER_SETTER(RGBAColor, color);
         GETTER_SETTER(float, interval);
         GETTER_SETTER(float, line_width);
-        GETTER_SETTER(int, max_vertices);
+        GETTER_SETTER(long, max_vertices);
 
         void gl_draw() override;
 };
